+++ src/index.ts
@@ -488,7 +488,7 @@ export type CreateQuery<T extends Record<string, unknown>> =
     (NonNullable<T> extends infer T
         ? {
               [K in keyof T]: T[K] extends (_: infer Params) => infer Query
-                  ? UnwrapArray<Query> extends Record<string, unknown>
+                  ? UnwrapArray<NonNullable<Query>> extends Record<string, unknown>
                       ? {
                             select: CreateQuery<UnwrapArray<Query>>
                             where: Params